ai_engine/requirements.txt
Flask
flask-cors
transformers
torch
sentence-transformers
faiss-cpu
numpy
pdfplumber
requests
spacy
lifelines
scikit-learn
pandas
google-generativeai
google-genai
python-dotenv
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_sm-0.5.4.tar.gz